Update README.md
I think this is better because it guides people to do what they really need. I was a bit lost because first I set the buildpack on my app to yours and my node app stopped to build. Then after reading more docs I understood that what I actually had to do was to add another buildpack, and not set the current one.